pyproject.toml
[project]
name = "agenthub"
version = "0.1.0"
description = "Production-ready AI Agent with FastAPI and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web 框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # 数据库
    "sqlmodel>=0.0.24",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.0",
    # AI/LLM - Ollama
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    # 长期记忆
    "mem0ai>=0.1.0",
    # 认证
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # 工具
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "asgiref>=3.8.0",
    "email-validator>=2.3.0",
    "bcrypt==4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
